requirements.txt
requests==2.32.3
pytest==8.3.4
python-dotenv==1.0.1
faker==33.3.1